Fix `lists` state change causing unnecessary unmounts
Previously, if the state of `lists` changes, all our editors would be destroyed. This could happen anytime someone added or removed a new list.

This is due to `lists` causing our initialization useEffect to rerun (it was in our useEffect dependency array), which causes the doc and providers to be destroyed and then inited again.

Issues spotted:
- When changing rooms and hot reloading, things don’t seem to load properly, but a fresh reload does it. Why?